+++ packages/moonstone/DatePicker/DatePickerBase.js
@@ -148,7 +148,7 @@ const DatePickerBase = kind({
 	render: ({day, maxDays, maxMonths, maxYear, minYear, month, noLabels, onChangeDate, onChangeMonth, onChangeYear, order, year, ...rest}) => {
 
 		return (
-			<ExpandableItemBase {...rest} showLabel="always">
+			<ExpandableItemBase {...rest} showLabel="always" autoCollpase={false} lockBottom={false}>
 				<div className={dateComponentPickers}>
 					{order.map(picker => {
 						switch (picker) {