+++ packages/moonstone/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `autoCollapse` and `lockBottom` properties to `ExpandableItem`
+
 ### Changed
 
 - `moonstone/Spinner` component's `center` and `middle` properties to a single `centered` property